# docs/swagger.yaml
basePath: /api/v1
definitions:
  helper.ErrorEnvelope:
    properties:
      error:
        type: string
    type: object
...
  payload.ProductCategoryRequest:
    properties:
      code:
        maxLength: 2
        minLength: 2
        type: string
      name:
        type: string
    required:
    - code
    - name
    type: object
  payload.ProductCategoryResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for the Golang RESTful API with Product, Book, Category,
    and ProductCategory CRUD operations
  title: Golang RESTful API
  version: 1.0.0
paths:
...
  /product-categories:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all product categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/payload.ProductCategoryResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorEnvelope'
      tags:
      - product-category
    post:
      consumes:
      - application/json
      description: Create a new product category with the provided details
      parameters:
      - description: Product category request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payload.ProductCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/payload.ProductCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorEnvelope'
      tags:
      - product-category
  /product-categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product category by its unique ID
      parameters:
      - description: Product Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorEnvelope'
      tags:
      - product-category
    get:
      consumes:
      - application/json
      description: Retrieve a single product category by its unique ID
      parameters:
      - description: Product Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.ProductCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorEnvelope'
      tags:
      - product-category
    put:
      consumes:
      - application/json
      description: Update an existing product category by its ID with the provided
        details
      parameters:
      - description: Product Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product category update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payload.ProductCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.ProductCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorEnvelope'
      tags:
      - product-category
...
swagger: "2.0"